+++ localstack-core/localstack/services/kinesis/plugins.py
@@ -1,8 +1,16 @@
+import localstack.config as config
 from localstack.packages import Package, package
 
 
 @package(name="kinesis-mock")
 def kinesismock_package() -> Package:
-    from localstack.services.kinesis.packages import kinesismock_package
+    from localstack.services.kinesis.packages import (
+        KinesisMockEngine,
+        kinesismock_package,
+        kinesismock_scala_package,
+    )
+
+    if KinesisMockEngine(config.KINESIS_MOCK_PROVIDER_ENGINE) == KinesisMockEngine.SCALA:
+        return kinesismock_scala_package
 
     return kinesismock_package
